+++ receiver/jaegerreceiver/trace_receiver_test.go
@@ -53,7 +53,6 @@ import (
 	"go.opentelemetry.io/collector/obsreport"
 	"go.opentelemetry.io/collector/testutil"
 	conventions "go.opentelemetry.io/collector/translator/conventions/v1.5.0"
-	tracetranslator "go.opentelemetry.io/collector/translator/trace"
 	"go.opentelemetry.io/collector/translator/trace/jaeger"
 )
 
@@ -355,8 +354,8 @@ func grpcFixture(t1 time.Time, d1, d2 time.Duration) *api_v2.PostSpansRequest {
 					StartTime:     t1,
 					Duration:      d1,
 					Tags: []model.KeyValue{
-						model.String(tracetranslator.TagStatusMsg, "Stale indices"),
-						model.Int64(tracetranslator.TagStatusCode, int64(pdata.StatusCodeError)),
+						model.String(conventions.OtelStatusDescription, "Stale indices"),
+						model.Int64(conventions.OtelStatusCode, int64(pdata.StatusCodeError)),
 						model.Bool("error", true),
 					},
 					References: []model.SpanRef{
@@ -374,8 +373,8 @@ func grpcFixture(t1 time.Time, d1, d2 time.Duration) *api_v2.PostSpansRequest {
 					StartTime:     t1.Add(d1),
 					Duration:      d2,
 					Tags: []model.KeyValue{
-						model.String(tracetranslator.TagStatusMsg, "Frontend crash"),
-						model.Int64(tracetranslator.TagStatusCode, int64(pdata.StatusCodeError)),
+						model.String(conventions.OtelStatusDescription, "Frontend crash"),
+						model.Int64(conventions.OtelStatusCode, int64(pdata.StatusCodeError)),
 						model.Bool("error", true),
 					},
 				},